demo_pneumonie/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
numpy
pillow
opencv-python-headless
scikit-image
matplotlib
huggingface_hub
